Adding support in USD for declaring properties as API schema overrides.
A schema property that is declared as an API schema override does not define
that property in that schema. Instead it behaves as an "over: for a property with
the same name that may be defined by an API schema that is included as a
built-in API schema of the schema declaring the override. This is different than
the behavior when included API schemas define the same property (without
specifying an override) as in that case, the strongest schema's property wins
and completely stomps over the same property from any weaker schema.

API schema overrides can be declared for both API and Typed schemas in the
source schema files through adding customData on the property of
'bool apiSchemaOverride = true'. This will be converted to a list of API schema
override property names stored in the schema prim's customData in the
resulting generatedSchema.usda. When the UsdSchemaRegistry is populated,
any properties declared as API schema overrides will not be immediately
added to the prim definition. Instead they will be composed over the defined
property from the schema's built-in API schemas if a property with that name
is indeed found. If no property with the name exists in the included API schemas,
the override property is ignored.

An examples of where this will be used is UsdLuxLightAPI which will now be able
to declare:
class "LightAPI" (
    prepend apiSchemas = ["CollectionAPI:lightLink", "CollectionAPI:shadowLink"]
) {
    uniform bool collection:lightLink:includeRoot = 1 (
        customData = {
            bool apiSchemaOverride = true
        }
    )
    uniform bool collection:shadowLink:includeRoot = 1 (
        customData = {
            bool apiSchemaOverride = true
        }
    )
}
This means that LightAPI can now override just the default values of the
includeRoot properties defined in its two built-in CollectionAPIs while leaving the
rest of the definition (such as documentation) as defined in the CollectionAPI
schema itself.

Another example is the light:shaderId property defined in LightAPI which all the
concrete Typed light schemas (RectLight, DistantLight, etc.) will be able to
override to defaulting to the USD shader ID for the light, once again, without
stomping over the entire property definition from LightAPI.

Note that this feature is specific to overriding properties from built-in API
schemas and doesn't apply to overriding properties via schema inheritance.
This is because schema properties are already always composed over via
inheritance during schema generation so no specification is necessary.

(Internal change: 2224838)