[HdSt] Updated Storm drawing coordinate plumbing
Updated interstage drawing coordinate plumbing to use resource
layouts. This facilitates using Hgi to manage interstage inputs
and outputs for Metal.

This also changes interstage drawing coord values to be independent
variables. Previously, drawing coord values were managed as members
of interstage struct values, but that will not work for Metal. An
alternative might be to use an interstage interface block, but that
is not robustly supported by cross platform GL drivers.

Since Hgi resource generation is not currently enabled for GL,
we process resource layouts in _GeneratateGLResources() when
compiling with generated GLSL resources. This is organized to
match the existing behavior when generating Hgi resources.

Also, while GLSL provides a way to pass array-valued input
attributes, we instead declare and bind input attributes
distinctly for each item in an array of instance indices which
better matches Metal.

Some of the pre-existing documentation for this is out-of-date,
and we will address that as a follow-up change.

Contribution; Thor Hjalmarsson, David G Yu

(Internal change: 2224924)