Add ability to add/remove/change Murk Sample Filters in usdview
Storing the sampleFilterPath indicated in the connection on
the RenderSettings prim in HdPrman's RenderParams, and
dirtying the affected sampleFilters when switching connected
filters. Then having the SampleFilter check against this stored
path to see if the filter should be created/deleted, and changing
the renderSettings to a Bprim so that it will be sure to sync
before the Sprim SampleFilter.

(Internal change: 2226882)