Physics: Add support for merging collision groups.
When merging assets from multiple sources, it's desirable in some
scenarios that we wish to merge physics collision groups. This adds
support for optional user-provided identifiers which allow such
merging. We also add an option to invert the selection of filtered
groups, which can be used to create colliders which opt-in to
collisions, rather than the default behaviour of having to opt-out.

Additionally, added a utility function which will calculate a table
indicating if two groups collide as well as tests for said utility. This
utility can be used in authoring use-cases to provide feedback for
content creators as well as providing a reference which can be used to
validate an implementation.