Let glslfxConfig parsing accept ints as a valid type.
Previously, it was impossible to consume an int-valued primvar in
a glslfx shader, as the config parsing in HioGlslfxConfig only handled
float, double, and vec thereof.  Now, it also handles ints (and thus
arrays of ints) but we do not bother extending to vec2i, vec3i, etc.

(Internal change: 2232543)