Remove "usd_" prefix from monolithic libraries by default.
Change 2190561 added the "usd_" prefix to all library
names by default, which caused the monolithic libraries
to be named "libusd_usd_ms.so" on Linux and macOS and
"usd_usd_ms.dll" on Windows. This changes the default
prefix for the monolithic builds so that the names
revert back to "libusd_ms.so" and "usd_ms.dll". Users
can still use the PXR_LIB_PREFIX option to change
the prefix if desired.

Issue and original proposed fix was from PR #1711

Fixes #1711

(Internal change: 2233803)