[hdSt/hdx] Add a render setting for the maximum number of lights storm should support.
Currently, storm is hardcoded to 16 lights via a single line in glf/simpleLightingContext.cpp.  Since we don't call into any API that explicitly restricts it to 16 lights, it's better to get rid of that line, and have SimpleLightTask truncate to a settings-provided number in a more clever manner.

Note that if you bump the number of lights past 16, you will encounter both performance issues (since we haven't optimized for this yet) and (possibly) issues with running out of uniform binding points, particularly for app integrations that use the lighting context to build their own lighting buffers instead of just using the hydra lighting buffers.  For now, playing with this setting should be considered experimental.

(Internal change: 2234026)