[hdx] Fix for GlfSimpleShadowArray::GetShadowMapTexture crash. Before updating GlfSimpleShadowArray with shadow information in HdxSimpleLightTask, check if lighting context will use shadows. Similarly, in HdxShadowTask, confirm there are a non-zero number of shadow maps in GlfSimpleShadowArray before creating renderpasses.
This avoids a discrepancy between the lighting context/light data and the
simple shadow array data. In the reported bug, there were 17 lights in the
lighting context, with only the last one possessing shadows. Because we only
support up to 16 lights in Storm, the last light is truncated (see
GlfSimpleLightingContext::SetLights). This resulted in a state in which the
lighting context did not think there were any shadows, while the simple shadow
array did.

(Internal change: 2234187)