Remove Maya-specific code for detecting Python from build_usd.py
This code was a remnant from when USD shipped with an optional
Maya plugin. Since that plugin has migrated to Autodesk's
maya-usd repository and USD itself no longer has Maya
dependencies, we can remove this code and reduce the surface
area of the (annoyingly complicated) Python detection code.

Note that the --build-python-info flag can be used to supply
Python information in cases where build_usd.py can't figure
it out itself. The documentation for this flag was updated
to clean it up a bit and remove explicit references to Maya.

(Internal change: 2235203)