tf: account for Python GIL explicit initialization deprecation from Python 3.7 to 3.9
Starting with Python 3.7, the GIL is initialized as part of Py_Initialize() and
PyEval_InitThreads() no longer needs to be called separately. In Python 3.9,
PyEval_InitThreads() no longer does anything and was marked deprecated.

See here for more detail:
https://docs.python.org/3/c-api/init.html#c.PyEval_InitThreads

This change avoids the use of PyEval_InitThreads() and
PyEval_ThreadsInitialized() when building against Python 3.7 or later.