Track ancestral dependencies introduced by culled nodes
Under certain composition structures (typically involving
inherit or specializes arcs), we ran into cases where
a node representing a new site dependency was introduced
at an ancestral level of namespace, but was then culled
from the prim index due to there being no specs further
down namespace. Pcp would fail to record this dependency
since it was not represented in the prim index. This
caused issues where Pcp would not report a change to the
prim index if an opinion was authored at that site later.

Pcp now records these "culled dependencies" in a separate
vector of PcpCulledDependency objects that is transferred
to Pcp_Dependencies at the end of prim indexing. These are
included by Pcp_Dependencies::ForEachDependencyOnSite and
Pcp_ForEachDependentNode. This required some refactoring
to share logic between dependencies represented by
PcpNodeRef and PcpCulledDependency objects.

All that said, it's possible that this issue could have
been fixed in a much simpler way by disabling culling
during the recursive prim indexing calls done in
_AddArc to pick up ancestral opinions. Conceptually
the idea is that the subtree we get back from this
recursive indexing represents a new set of dependencies
that have to stay in the prim index to mark those
dependencies, even if they are ancestral. Pcp already
has logic to avoid culling nodes that introduce
dependencies, it just isn't firing in this case.

I prototyped this simpler change and it did fix the
original problem. However, it exposed a pre-existing
issue in the museum case TrickySpookyVariantSelectionInClass
that is non-trivial to fix. The change I'm submitting
maintains the current (but broken) behavior, but once
that's fixed we may be able to come back and undo all of
this.

(Internal change: 2238230)
(Internal change: 2238320)