[hdSt] Incorrect MaterialX shading without Dome Light
MaterialX uses samplers in its shaderGen to calculate indirect
lighting. When there is a domeLight in the scene or the default
domeLight is enabled these are initialized with the appropriate
HdGetSampler_domeLight**() functions. If there is no domeLlight,
MaterialX still expects valid samplers. When using non-bindless
textures these samplers are declared without bindings specified,
and as a result if the MaterialX material has textures, the shader
will sample from those material textures for the indirect lighting.

Adding a fallback domeLight texture node to the material network
in MaterialXFilter, so that when there is no domeLight we will
sample from this fallback texture instead.

Fixes #1877

(Internal change: 2238260)