Let HdSt update BARs without changing the order of the buffers therein, whenever possible.
This fixes an excessive-shader-recompilation bug in which an existing
mesh drawItem, when updated, could, while trying to sparsely update,
rearrange some of its primvars (in this case, points and
packedSmoothNormals). The resulting shader recompilation request
would trigger a shader cache miss and cause costly recompilation,
perceived as a "hitch" in drawing.

(Internal change: 2240666)