Fix inconsistent property search result order in usdview
The order in which properties were highlighted when using
the "Find Prop" search bar was inconsistent. When there
was no exact match for the search string, the results
were presented from top-to-bottom of the view. But if
there *was* an exact match, that property would be
highlighted first, followed by all other matches that
contained the search string.

This change makes the behavior consistent by sorting the
search results by their model index, so that results are
always presented from top-to-bottom of the view. This
also fixes an issue where the order would differ between
PySide2 and PySide6 due to the use of QRegExp in PySide2
vs QRegularExpression in PySide6.

(Internal change: 2241157)