[usdMtlx] Add support for locally defined custom nodes
Locally defined custom nodes are those defined either alongside
and in the same file as the material or in a separate MaterialX
file that is then included at the top of the the material.mtlx file.
UsdMtlx will flatten this custom node out and essentially replace
it with it's associated nodegraph. As such, support for these
custom nodes are currently limited to those defined with a
nodegraph as opposed to a language specific implementation.

Fixes #1636

(Internal change: 2242185)