Fix bug where deleting a variant set containing only empty variants would try to delete the variants twice, leading to errors the second time around.
SdfLayer::_DeleteSpec previously deleted specs in an inert
subtree one at a time from the bottom up, so that change
notices for each inert spec would be registered. The
_PrimDeleteSpec function it uses recurses over entire subtrees,
so _DeleteSpec was careful to clear out the children fields
on prims to avoid recursing over specs that were already deleted.
However, this didn't happen for variant sets, which caused
the error.

_DeleteSpec now traverses inert subtrees to register the
desired change notices before calling _PrimDeleteSpec just
once to delete the entire subtree. This is simpler and more
efficient: we no longer need to collect the paths of each
spec being deleted so we can do the bottom-up traversal, and
we don't need to modify fields on specs that are about to
be deleted anyway.

(Internal change: 2245070)