[HdSt, HdMtlx] Add mtlx shader to the resourceRegistry for simple MaterialX materials
For simple materials, ones with just a terminal node,
add the generated ShaderPtr to the resource registry
for future simple materials to re-use.
Note that transparent materials have a different material
tag, which is stored in the glslfx header of the ShaderPtr,
and StandardSurface transparant materials have a slightly
different shader so shaders cannot be re-used between
transparent and opaque materials.

(Internal change: 2246342)