[usdGeom] Edit gathering of geom subset child prims to not use UsdPrimDefaultPredicate, and instead use a custom predicate that checks if the prim has a defining specifier, rather than is defined.
This is done to allow for correct gathering of geom susbets when the geom prim
is parented by a not-defined prim; the pertinent case being being when the geom
prim is a child of a "Prototype" prim as part of a PointInstancer setup. The
"Prototype" prim is often specified using "over", which can lead to that prim
and its children not being "defined".

(Internal change: 2246474)