[usdMtlx, hdMtlx, HdSt] Adding support for Custom Material Nodes that use textures.
Note that when using textures custom material nodes need to
expose the texture, meaning the 'filename' must be an input
on the material. Also, they need to be defined in a shared
library so the corresponding sdrNodes are created.

Adding surfaceshader entry to table in UsdMtlxUtils to fix a
warning when using custom material nodes.

Fixes #1786

(Internal change: 2248719)