Separating UsdAttribute::GetResolveInfo(UsdTimeCode time = UsdTimeCode::Default()) into two separate overloads: UsdAttribute::GetResolveInfo(UsdTimeCode time) UsdAttribute::GetResolveInfo() This is in preparation for fixing calls to GetResolveInfo with default time which erroneously considers time samples when time samples aren't resolved at default time. The separate overload of GetResolveInfo() is added to maintain the existing behavior of calling GetResolveInfo without a time which now does resolve time samples (but ignores clips).
Note that this change doesn't functionally change anything and the
behavior of calling GetResolveInfo at default time is not actually
fixed yet. This change is separated out because the mere addition of
this overload with no functional change is causing some startup time
performance regressions in gauntlet runs and it'll be useful to pinpoint
those to this, keeping them out of the meatier upcoming attribute resolve
changes.

(Internal change: 2248911)