Unify behavior of procedural resolving scene index's GetChildPrimPaths to match PrimsAdded forwarding of existing child prims
_PrimsAdded does not filter out existing descendents of procedurals. Previously, GetChildPrimPaths did not report
them beneath any scope for which a procedural was adding children. That meant that there was a
mismatch in what was present based on whether you relied on notices or direct traversal in order
to interpret the scene. It was (and remains) possible for a procedural to mask an input descendent
in order to override its type or data sources. But if no action is taken by the procedural, the descendent prim's
presence is now consistently reported.

(Internal change: 2249457)