(crash issues in testVP2RenderDelegateUsdSkel)
The problem here was failure while invoking the skinning computation
due to an incorrect check for a computation input that was newly added.
Basically the 'skinningMethod' input could sometimes have an empty value
(and a fallback was being provided at the time of access) but there was a
check before for holding TfToken type which failed and raised a computation error
to the HdExtComputationContext.
This was causing the test to crash with no stack.
Fixing by removing the unnecessary check.

(Internal change: 2249802)