[UsdImaging] Fixing failed verify in HdPrman when processing CoordSys
The delegateId was being pre-emptively added as a prefix to
the coordSys prim paths and then additionally added in the
IndexProxyAdapter. This double adding meant that we were
unable to find the coordSys prim in HdPrman_RenderParam
and caused the validation error. This change removes this
pre-emptive adding of the delegateId prefix and instead adds
the prefix before retrning the path vector in
UsdImagingDelegate::GetCoordSysBindings().

Usd test files are based on the files in testHdPrman_CoordSys

(Internal change: 2249854)