Updated USD azure pipelines CI to use Python 3
This is a follow up to the recent change which updated
the azure pipelines agents for Ubuntu and macOS.

With this change we're also updating the macOS build to
use PySide6 and the Windows build to use PySide2.

Here's a summary of the updated configurations:

  - Ubuntu 20.04, Python 3.8.10, PySide2 5.15.2.1
  - macOS 11.7, Python 3.10.6, PySide6 6.3.2
  - Windows 10, Visual Studio 2019, Python 3.7.9, PySide2 5.15.2.1

Also, increased the timeout from 90 to 120 minutes since we are
currently seeing build times of arround 80-110 minutes.

(Internal change: 2250003)