Fixing a performance regression.
It appears that there is a non-trivial cost to the default
initialization of the PcpNodeIterator and the layer vector iterator
added to Usd_Resolver to support early iteration stopping for
UsdResolveTargets. This cost can be alleviated by just storing a pointer
to the resolve target itself as the resolve target already holds the
iterators needed. The commonly constructed Usd_Resolver now just has to
initialize this pointer to nullptr instead of holding the two iterators.

(Internal change: 2250889)