OGSMOD-8086: More robust clip plane shader data.
Make sure clip planes are properly reset when they are enabled and then disabled with HdRenderPassState::SetClippingEnabled(bool enabled).
Note: the same shader gen code is used for clip planes count between 0-4, even when clipping is disabled.
Consequently, we need to update the clip plane count data, so the shader has its input buffers updated and stops applying the previous clip planes value when the feature is disabled.

The only safe scenario where we can skip setting the clip plane buffer source is when the clip feature is not supported.